refactor: payments 관련 내용 메소드 위치 옮기기
