+++ src/main/java/com/timeToast/timeToast/service/icon/AdminIconService.java
@@ -29,7 +29,4 @@ public interface AdminIconService {
 
     IconGroupAdminResponses getAllIconGroups();
 
-    IconGroupMonthlyRevenues iconGroupMonthlyRevenue(final int year);
-
-
 }
